Module psutil::process

Read process-specific information from /proc

More information about specific fields can be found in proc(5).

Field sizes

The manual pages for proc define integer sizes using scanf(3) format specifiers, which parse to implementation specific sizes. This is obviously a terrible idea, and so this code makes some assumptions about the sizes of those specifiers.

These assumptions are backed up by libc::types::os::arch::posix88::pid_t, which declares PIDs to be signed 32 bit integers. proc(5) declares that PIDs use the %d format specifier.

%d / %u - 32 bit signed and unsigned integers

%ld / %lu - 64 bit signed and unsigned integers

WARNING: Rust currently has no support for 128 bit integersrfc521 so %llu (used by the starttime and delayacct_blkio_ticks fields) is is instead represented by a 64 bit integer, with the hope that doesn't break.

CPU time fields and clock ticks

The CPU time fields are very strange. Inside the Linux kernel they each use the same typearray.c:361 but when printed use different typesarray.c:456 - the fields utime, stime and gtime are unsigned integers, whereas cutime, cstime and cgtime are signed integers.

These values are all returned as a number of clock ticks, which can be divided by sysconf(_SC_CLK_TCK) to get a value in seconds. The Process struct does this conversion automatically, and all CPU time fields use the f64 type.

Structs

Memory
Memory usage of a process

Process
Information about a process gathered from /proc/[pid]/stat.

Enums

State
Possible statuses for a process

Functions

all
Return a vector of all processes in /proc