Struct psutil::process::Process

pub struct Process {
    pub pid: PID,
    pub comm: String,
    pub state: State,
    pub ppid: PID,
    pub pgrp: i32,
    pub session: i32,
    pub tty_nr: i32,
    pub tpgid: i32,
    pub flags: u32,
    pub minflt: u64,
    pub cminflt: u64,
    pub majflt: u64,
    pub cmajflt: u64,
    pub utime: f64,
    pub stime: f64,
    pub cutime: f64,
    pub cstime: f64,
    pub priority: i64,
    pub nice: i64,
    pub num_threads: i64,
    pub starttime: u64,
    pub vsize: u64,
    pub rss: i64,
    pub rsslim: u64,
    pub wchan: u64,
    pub exit_signal: i32,
    pub processor: i32,
    pub rt_priority: u32,
    pub policy: u32,
    pub delayacct_blkio_ticks: u64,
    pub guest_time: f64,
    pub cguest_time: f64,
    pub exit_code: i32,
    // some fields omitted
}

Information about a process gathered from /proc/[pid]/stat.

IMPORTANT: See the module level notes for information on the types used by this struct, as some do not match those used by /proc/[pid]/stat.

Fields

pid

PID of the process

comm

Filename of the executable

state

State of the process as an enum

ppid

PID of the parent process

pgrp

Process group ID

session

Session ID

tty_nr

Controlling terminal of the process [TODO: Actually two numbers]

tpgid

ID of the foreground group of the controlling terminal

flags

Kernel flags for the process

minflt

Minor faults

cminflt

Minor faults by child processes

majflt

Major faults

cmajflt

Major faults by child processes

utime

Time scheduled in user mode (seconds)

stime

Time scheduled in kernel mode (seconds)

cutime

Time waited-for child processes were scheduled in user mode (seconds)

cstime

Time waited-for child processes were scheduled in kernel mode (seconds)

priority

Priority value (-100..-2 | 0..39)

nice

Nice value (-20..19)

num_threads

Number of threads in the process

starttime

Time the process was started after system boot (clock ticks)

vsize

Virtual memory size in bytes

rss

Resident Set Size (pages) [TODO: Calculate size in bytes]

rsslim

Current soft limit on process RSS (bytes)

wchan

Channel the process is waiting on (address of a system call)

exit_signal

Signal sent to parent when process dies

processor

Number of the CPU the process was last executed on

rt_priority

Real-time scheduling priority (0 | 1..99)

policy

Scheduling policy

delayacct_blkio_ticks

Aggregated block I/O delays (clock ticks)

guest_time

Guest time of the process (seconds)

cguest_time

Guest time of the process's children (seconds)

exit_code

The thread's exit status

Methods

impl Process

fn new(pid: PID) -> Result<Process>

Attempts to read process information from /proc/[pid]/stat.

/stat is seperated by spaces and contains a trailing newline.

This should return a psutil/process specific error type, so that errors can be raised by FromStr too

fn from_pidfile(path: &Path) -> Result<Process>

Create a Process by reading it's PID from a pidfile.

fn is_alive(&self) -> bool

Return true if the process was alive at the time it was read.

fn cmdline_vec(&self) -> Result<Option<Vec<String>>>

Read /proc/[pid]/cmdline as a vector.

Returns Err if /proc/[pid]/cmdline is empty.

fn cmdline(&self) -> Result<Option<String>>

Return the result of cmdline_vec as a String.

fn memory(&self) -> Result<Memory>

Reads /proc/[pid]/statm into a struct.

fn kill(&self) -> Result<()>

Send SIGKILL to the process.